Drag and drop
- images
- vidéo
- pdf
I added a preview of the items below the drag and drop field
I added 2 buttons in the preview part which allow the user to remove or rename his photo/document
I added a function in the js that rename automaticaly the images files into the right name for the template to work with them.
I added a logo part in which the user can provide his logo and we'll use it for the generation
I added a color analytic in the code that retreive the rgb code of the main color from the image and apply to the buttons in the preview (still working on it)
I added the link with the database
And I added the possibility to use the code part as well for the client and as an intern use.
Des bisous <3